schema update

+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/autodiff/samediff/SameDiff.java
@@ -6151,7 +6151,7 @@ public class SameDiff {
             extras[e] = ((Number) node.getExtraArgs()[e]).floatValue();
         }
 
-        int[] extraBits = null;
+        long[] extraBits = null;
         if (node.opType() == Op.Type.CUSTOM) {
             DynamicCustomOp dynamicCustomOp = (DynamicCustomOp) node;
             extraBits = dynamicCustomOp.iArgs();
@@ -6161,9 +6161,9 @@ public class SameDiff {
             if (!framesMap.containsKey(frameName))
                 framesMap.put(frameName, idCounter.incrementAndGet());
 
-            extraBits = new int[]{framesMap.get(frameName).intValue()};
+            extraBits = new long[]{framesMap.get(frameName).intValue()};
         } else
-            extraBits = new int[]{};
+            extraBits = new long[]{};
 
         val inPaired = new ArrayList<Integer>();
